chore() change folder path and rename dart file to match project
