+++ lib/src/apivideo_player_overlay.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'dart:math';
 
-import 'package:apivideo_player/presentation/api_video_icons_icons.dart';
+import 'package:apivideo_player/src/presentation/apivideo_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:pointer_interceptor/pointer_interceptor.dart';
