+++ pubspec.yaml
@@ -57,7 +57,7 @@ flutter:
   fonts:
     - family:  ApiVideoIcons
       fonts:
-       - asset: fonts/ApiVideoIcons.ttf
+       - asset: lib/src/fonts/ApiVideoIcons.ttf
   # To add assets to your plugin package, add an assets section, like this:
   # assets:
   #   - images/a_dot_burr.jpeg
